Move Params class's to Protocol project
