+++ src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer/EndpointContracts/Debugging/IRazorProximityExpressionsEndpoint.cs
@@ -2,6 +2,7 @@
 // Licensed under the MIT license. See License.txt in the project root for license information.
 
 using Microsoft.AspNetCore.Razor.LanguageServer.Common;
+using Microsoft.AspNetCore.Razor.LanguageServer.Protocol;
 using OmniSharp.Extensions.JsonRpc;
 
 namespace Microsoft.AspNetCore.Razor.LanguageServer.EndpointContracts.Debugging;
