+++ src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer/EndpointContracts/IRazorBreakpointSpanEndpoint.cs
@@ -2,11 +2,12 @@
 // Licensed under the MIT license. See License.txt in the project root for license information.
 
 using Microsoft.AspNetCore.Razor.LanguageServer.Common;
+using Microsoft.AspNetCore.Razor.LanguageServer.Protocol;
 using OmniSharp.Extensions.JsonRpc;
 
 namespace Microsoft.AspNetCore.Razor.LanguageServer.EndpointContracts;
 
 [Parallel, Method(LanguageServerConstants.RazorBreakpointSpanEndpoint)]
-internal interface IRazorBreakpointSpanEndpoint : IJsonRpcRequestHandler<RazorBreakpointSpanParams, RazorBreakpointSpanResponse?>
+internal interface IRazorBreakpointSpanEndpoint : IJsonRpcRequestHandler<RazorBreakpointSpanParamsBridge, RazorBreakpointSpanResponse?>
 {
 }
