+++ src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer/EndpointContracts/RazorBreakpointSpanParamsBridge.cs
@@ -1,0 +1,10 @@
+﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the MIT license. See License.txt in the project root for license information.
+
+using MediatR;
+
+namespace Microsoft.AspNetCore.Razor.LanguageServer.Protocol;
+
+internal class RazorBreakpointSpanParamsBridge : RazorBreakpointSpanParams, IRequest<RazorBreakpointSpanResponse>
+{
+}
